+++ environment.yml
@@ -15,3 +15,6 @@ dependencies:
     - python-dotenv
     - roboflow
     - opencv-python-headless>=4.10.0
+    - protobuf
+    - asyncio
+    - opencv-python
